Move RTM to own file
