+++ package.json
@@ -25,6 +25,8 @@
     "dotenv": "^5.0.1"
   },
   "devDependencies": {
+    "babel-preset-env": "^1.7.0",
+    "babel-register": "^6.26.0",
     "nodemon": "^1.17.5"
   }
 }
